restructured, components are self contained within folders
