+++ src/components/AboutPage/AboutPage.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
-import '../styles/about-page.css'
+import './AboutPage.css'
 
 // Since this component is simple and static, there's no parent container for it.
 const AboutPage = () => {
@@ -8,8 +8,7 @@ const AboutPage = () => {
     <div>
       <h2 className='alt-header'>About</h2>
       <p>
-        This example app is part of the <a href='https://github.com/coryhouse/react-slingshot'>React-Slingshot
-        starter kit</a>.
+        Basic starter repo
       </p>
       <p>
         <Link to='/badlink'>Click this bad link</Link> to see the 404 page.
