+++ src/index.js
@@ -4,10 +4,14 @@ import React from 'react'
 import { render } from 'react-dom'
 import { AppContainer } from 'react-hot-loader'
 import { BrowserRouter } from 'react-router-dom'
+
 import configureStore, { history } from './store/configureStore'
 import Root from './components/Root'
-import './styles/styles.scss' // Yep, that's right. You can import SASS/CSS files too! Webpack will run the associated loader and plug this into the page.
+
+import './index.scss'
+
 require('./favicon.ico') // Tell webpack to load favicon.ico
+
 const store = configureStore()
 
 render(
