+++ webpack.config.dev.js
@@ -98,24 +98,16 @@ export default {
         use: [
           'style-loader',
           {
-            loader: 'css-loader',
-            options: {
-              sourceMap: true
-            }
+            loader: 'css-loader'
           }, {
             loader: 'postcss-loader',
             options: {
               plugins: () => [
                 require('autoprefixer')
-              ],
-              sourceMap: true
+              ]
             }
           }, {
-            loader: 'sass-loader',
-            options: {
-              includePaths: [path.resolve(__dirname, 'src', 'scss')],
-              sourceMap: true
-            }
+            loader: 'sass-loader'
           }
         ]
       }
